Safari-609.1.11.1/LayoutTests/fast/html/link-element-removal-during-beforeload-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: runTest
This tests removing a node during beforeload event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("link") is exitingLink
PASS successfullyParsed is true

TEST COMPLETE
